chore!: modernize plugin
1. JS -> TS
2. update to React 18 and favor hooks over HOC
3. update to studio v3 API

This change rewrites the plugin to accomplish the above goals.  This is still non-functional overall though as it does not update the deprecated and sunset Google Analytics API currently in use.

BREAKING CHANGE: this change requires users to pass their client ID and views ID as part of the plugin config declaration rather than stored in a json file.

BREAKING CHANGE: the `withAnalyticsData` HOC has been replaced by the GoogleAnalyticsReportProvider context.